Move localPtyService to node
Part of #116467
