+++ src/vs/platform/terminal/node/localPtyService.ts
